[13-navigation-menu-with-categories] Separated reducer from context.
